REFACTOR: Moved AccordionItem to new file
